file restructure/ code reformat to MVC
